Polish Rooms CRUD and Styling
- More view types in room detail page
- Multiple image support
- Merge and improve add and edit pages
- Streamline room list page

+++ lib/services/firestore_service.dart
@@ -1,6 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter/cupertino.dart';
 
 import '../model/room_data.dart';
+import '../utils/ui_utils.dart';
 
 class FirestoreService {
   final CollectionReference rooms =
@@ -18,45 +20,25 @@ class FirestoreService {
   }
 
   // UPDATE
-  Future<void> updateRoom(String docID, RoomData room) {
-    return rooms.doc(docID).update(room.toFirestore());
+  Future<void> updateRoom(RoomData room, BuildContext context) async {
+    var docID = room.id;
+    final confirmed = await UiUtils.showPrompt(
+      context,
+      'Are you sure you want to update this room?',
+    );
+    if (confirmed?? false) {
+      return rooms.doc(docID).update(room.toFirestore());
+    }
   }
 
   // DELETE
-  Future<void> deleteRoom(String docID) {
-    return rooms.doc(docID).delete();
-  }
-
-  Future<bool> roomsExist() async {
-    final snapshot = await rooms.limit(1).get();
-    return snapshot.docs.isNotEmpty;
-  }
-
-  Future<RoomData> getInitialRoom() async {
-    final querySnapshot = await rooms.limit(1).get();
-    if (querySnapshot.docs.isNotEmpty) {
-      return RoomData.fromFirestore(querySnapshot.docs.first);
-    } else {
-      // If no rooms exist, return a default room
-      RoomData initialRoom =  RoomData(
-        id: '1',
-        name: 'Kamar 1',
-        description:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-        price: '1.500.000',
-        roomImages: [
-          RoomImage(imageLink: 'https://via.placeholder.com/400'),
-          RoomImage(imageLink: 'https://via.placeholder.com/401'),
-        ],
-        roomFacilities: [
-          Facility(name: 'Free Wi-Fi'),
-          Facility(name: 'Air Conditioning'),
-        ],
-        availability: true,
-      );
-
-      await addRoom(initialRoom);
-      return initialRoom;
+  Future<void> deleteRoom(String docID, BuildContext context) async {
+    final confirmed = await UiUtils.showPrompt(
+      context,
+      'Are you sure you want to delete this room?',
+    );
+    if (confirmed?? false) {
+      return rooms.doc(docID).delete();
     }
   }
 }
